+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Adds the custom Activity to call DDB Macro functions and World Macros.
 - Darkness spell effect using said activity.
+- Spell Enchantment and Summoning Effects.
 
 # 6.0.3